Move config into a country specific folder
