+++ public/data/mongolia/groundstations/IRIMHE_precipitation_2020-05-01.json
