+++ public/data/mongolia/groundstations/IRIMHE_temperature_2020-01-d1-trim.json
