+++ public/data/mongolia/groundstations/longterm_data.json
