+++ public/data/mongolia/nso/NSO_Disabled_Admin1_Male.json
