+++ public/data/mongolia/nso/NSO_Herd_Size_Admin1_LT_200.json
