+++ public/data/mongolia/nso/NSO_Herder_HHs_ths_Admin2.json
