+++ public/data/mongolia/tables/NSO_Animal_Deaths_Annual_Admin1.csv
