+++ public/data/mongolia/tables/NSO_Bale_Hay_Price_Monthly_Admin1.csv
