+++ public/data/mongolia/tables/NSO_Bale_Hay_Price_Monthly_Avg2.csv
