+++ public/data/mongolia/tables/wfp_food_prices_mongolia.csv
