+++ src/config/config.ts
@@ -1,0 +1,5 @@
+import appConfig from './mongolia/prism.json';
+import rawLayers from './mongolia/layers.json';
+import rawTables from './mongolia/tables.json';
+
+export { appConfig, rawLayers, rawTables };
