+++ src/config/mongolia/tables.json
@@ -1,7 +1,7 @@
 {
   "animal_deaths": {
     "title": "Animal deaths due to disease or natural hazards",
-    "table": "./data/tables/NSO_Animal_Deaths_Annual_Admin1.csv",
+    "table": "../data/mongolia/tables/NSO_Animal_Deaths_Annual_Admin1.csv",
     "legend_text": "Annual deaths of animals due to disease or natural hazards (including dzud). Source: National Statistics Office",
     "chart": {
       "type": "line",
@@ -13,7 +13,7 @@
   },
   "hay_prices": {
     "title": "Monthly hay prices",
-    "table": "./data/tables/NSO_Bale_Hay_Price_Monthly_Avg2.csv",
+    "table": "../data/mongolia/tables/NSO_Bale_Hay_Price_Monthly_Avg2.csv",
     "legend_text": "Price of a bail of hay by month. Source: National Statistics Office",
     "chart": {
       "type": "line",
@@ -25,7 +25,7 @@
   },
   "poverty": {
     "title": "Poverty headcount 2016 vs 2018",
-    "table": "./data/tables/NSO_Poverty_Headcount_Admin1.csv",
+    "table": "../data/mongolia/tables/NSO_Poverty_Headcount_Admin1.csv",
     "legend_text": "Poverty headcount over two periods. Source: National Statistics Office",
     "chart": {
       "type": "bar",
